Refactor and rename cart table component
